The module was rewritten according to ECMAScript notation and renamed to esm.mjs.
